Change column width calculation logic (#747)
Previously, the viewport width is divided between available columns. If some cell has width inline styling, it will be applied directly which may cause weird rendering issue (see #733).

In this PR, the cell intrinsic width is taken into account but not used directly to make sure table contents are all visible. Because of this logic, table cells may be rendered outside of the reported canvas size. Wrapping `HtmlTable` within some `ClipRect` should help if that behavior is undesired.

Related to #531